Uploading SES project to Github
This is phase one of SES. This phase is covering components of simple AI Scaling. There are U.I elements/features such as: Main Menu, Settings Menu. Game play features such as a talent tree setup and pause menu. There are also in-game UI features that showcase player health and stats as the player levels up.